Updating PublishWindow to match the style of AuthenticationWindow
