+++ src/UnityExtension/Assets/Editor/GitHub.Unity/GitHub.Unity.csproj
@@ -84,6 +84,7 @@
     <Compile Include="Logging\UnityLogAdapter.cs" />
     <Compile Include="EntryPoint.cs" />
     <Compile Include="Misc\Installer.cs" />
+    <Compile Include="UI\PublishWindow.cs" />
     <Compile Include="UI\ProjectWindowInterface.cs" />
     <Compile Include="Misc\Styles.cs" />
     <Compile Include="Misc\Utility.cs" />
@@ -100,7 +101,7 @@
     <Compile Include="UI\GitCommitTarget.cs" />
     <Compile Include="UI\HistoryView.cs" />
     <Compile Include="UI\IView.cs" />
-    <Compile Include="UI\PublishWindow.cs" />
+    <Compile Include="UI\PublishView.cs" />
     <Compile Include="UI\SettingsView.cs" />
     <Compile Include="UI\Subview.cs" />
     <Compile Include="UI\Window.cs" />
@@ -112,7 +113,7 @@
   </ItemGroup>
   <Choose>
     <When Condition="$(Buildtype) == 'Internal'">
-      <ItemGroup >
+      <ItemGroup>
         <Compile Include="$(SolutionDir)\script\src\MetricsInitialization.cs">
           <Link>Metrics\MetricsInitialization.cs</Link>
         </Compile>
