+++ src/UnityExtension/Assets/Editor/GitHub.Unity/UI/PublishWindow.cs
@@ -1,247 +1,76 @@
-﻿using System;
-using System.Linq;
-using Octokit;
+using System;
 using UnityEditor;
 using UnityEngine;
 
 namespace GitHub.Unity
 {
+    [Serializable]
     class PublishWindow : BaseWindow
     {
-        private const string WindowTitle = "Publish";
-        private const string Title = "Publish this repository to GitHub";
-        private const string PrivateRepoMessage = "You choose who can see and commit to this repository";
-        private const string PublicRepoMessage = "Anyone can see this repository. You choose who can commit";
-        private string repoName = "";
-        private string repoDescription = "";
-        private int selectedOrg = 0;
-        private bool togglePrivate = false;
-        private bool isBusy = false;
-        private string error;
+        private const string Title = "Publish";
 
-        private string username;
-        private string[] owners = { };
-        private IApiClient client;
+        [SerializeField] private PublishView publishView;
 
-        public IApiClient Client
+        public static void Launch()
         {
-            get
-            {
-                if (client == null)
-                {
-                    var repository = Environment.Repository;
-                    UriString host;
-                    if (repository != null && !string.IsNullOrEmpty(repository.CloneUrl))
-                    {
-                        host = repository.CloneUrl.ToRepositoryUrl();
-                    }
-                    else
-                    {
-                        host = UriString.ToUriString(HostAddress.GitHubDotComHostAddress.WebUri);
-                    }
-
-                    client = ApiClient.Create(host, Platform.Keychain);
-                }
-
-                return client;
-            }
+            Open();
         }
 
         public static IView Open(Action<bool> onClose = null)
         {
-            var publishWindow = GetWindow<PublishWindow>();
-
+            PublishWindow publishWindow = GetWindow<PublishWindow>(true);
             if (onClose != null)
                 publishWindow.OnClose += onClose;
-
-            publishWindow.minSize = new Vector2(300, 250);
+            publishWindow.minSize = publishWindow.maxSize = new Vector2(300, 250);
             publishWindow.Show();
-
             return publishWindow;
+       }
+
+        public override void OnUI()
+        {
+            if (publishView == null)
+            {
+                CreateViews();
+            }
+            publishView.OnGUI();
+        }
+
+        public override void Refresh()
+        {
+            publishView.Refresh();
         }
 
         public override void OnEnable()
         {
             // Set window title
-            titleContent = new GUIContent(WindowTitle, Styles.SmallLogo);
+            titleContent = new GUIContent(Title, Styles.SmallLogo);
 
-            Utility.UnregisterReadyCallback(PopulateView);
-            Utility.RegisterReadyCallback(PopulateView);
+            Utility.UnregisterReadyCallback(CreateViews);
+            Utility.RegisterReadyCallback(CreateViews);
+
+            Utility.UnregisterReadyCallback(ShowActiveView);
+            Utility.RegisterReadyCallback(ShowActiveView);
         }
 
-        private void PopulateView()
+        private void CreateViews()
         {
-            try
-            {
-                Initialize(EntryPoint.ApplicationManager);
+            if (publishView == null)
+                publishView = new PublishView();
 
-                var keychainConnections = Platform.Keychain.Connections;
-                if (keychainConnections.Any())
-                {
-                    username = keychainConnections[0].Username;
-                    owners = new[] { username };
-                }
-                else
-                {
-                    Logger.Warning("No Keychain connections to use");
-                }
-
-                Logger.Trace("Create Client");
-
-                Logger.Trace("GetOrganizations");
-
-                Guard.NotNull(this, Platform, "Platform");
-                Guard.NotNull(Platform, Platform.Keychain, "Platform.Keychain");
-
-                Client.GetOrganizations(organizations => {
-                    if (organizations == null)
-                    {
-                        Logger.Warning("Organizations is null");
-                        return;
-                    }
-
-                    Logger.Trace("Loaded {0} organizations", organizations.Count);
-
-                    var organizationLogins = organizations
-                        .OrderBy(organization => organization.Login)
-                        .Select(organization => organization.Login);
-
-                    owners = owners.Union(organizationLogins).ToArray();
-                });
-
-            }
-            catch (Exception e)
-            {
-                Logger.Error(e, "Error PopulateView & GetOrganizations");
-                throw;
-            }
+            Initialize(EntryPoint.ApplicationManager);
+            publishView.InitializeView(this);
         }
 
-        public override void OnGUI()
+        private void ShowActiveView()
         {
-            base.OnGUI();
+            publishView.OnShow();
+            Refresh();
+        }
 
-            GUILayout.BeginHorizontal(Styles.AuthHeaderBoxStyle);
-            {
-                GUILayout.BeginVertical(GUILayout.Width(16));
-                {
-                    GUILayout.Space(9);
-                    GUILayout.Label(Styles.BigLogo, GUILayout.Height(20), GUILayout.Width(20));
-                }
-                GUILayout.EndVertical();
-
-                GUILayout.BeginVertical();
-                {
-                    GUILayout.Space(11);
-                    GUILayout.Label(Title, EditorStyles.boldLabel); 
-                }
-                GUILayout.EndVertical();
-            }
-            GUILayout.EndHorizontal();
-
-            GUILayout.Space(5);
-
-            GUILayout.BeginHorizontal();
-            {
-                GUILayout.BeginVertical();
-                {
-                    GUILayout.Label("Owner");
-
-                    GUI.enabled = !isBusy;
-                    selectedOrg = EditorGUILayout.Popup(0, owners);
-                    GUI.enabled = true;
-                }
-                GUILayout.EndVertical();
-
-                GUILayout.BeginVertical(GUILayout.Width(8));
-                {
-                    GUILayout.Space(20);
-                    GUILayout.Label("/");
-                }
-                GUILayout.EndVertical();
-
-                GUILayout.BeginVertical();
-                {
-                    GUILayout.Label("Repository Name");
-                    GUI.enabled = !isBusy;
-                    repoName = EditorGUILayout.TextField(repoName);
-                    GUI.enabled = true;
-                }
-                GUILayout.EndVertical();
-            }
-            GUILayout.EndHorizontal();
-
-            GUILayout.Label("Description");
-            GUI.enabled = !isBusy;
-            repoDescription = EditorGUILayout.TextField(repoDescription);
-            GUI.enabled = true;
-            GUILayout.Space(5);
-
-            GUILayout.BeginVertical();
-            {
-                GUILayout.BeginHorizontal();
-                {
-                    GUI.enabled = !isBusy;
-                    togglePrivate = GUILayout.Toggle(togglePrivate, "Create as a private repository");
-                    GUI.enabled = true;
-                }
-                GUILayout.EndHorizontal();
-
-                GUILayout.BeginHorizontal();
-                {
-                    GUILayout.Space(5);
-                    var repoPrivacyExplanation = togglePrivate ? PrivateRepoMessage : PublicRepoMessage;
-                    GUILayout.Label(repoPrivacyExplanation, Styles.LongMessageStyle);
-                }
-                GUILayout.EndHorizontal();
-            }
-            GUILayout.EndVertical();
-
-
-            GUILayout.Space(5);
-
-            if (error != null)
-                GUILayout.Label(error, Styles.ErrorLabel);
-
-            GUILayout.FlexibleSpace();
-
-            GUILayout.BeginHorizontal();
-            {
-                GUILayout.FlexibleSpace();
-                GUI.enabled = !string.IsNullOrEmpty(repoName) && !isBusy;
-                if (GUILayout.Button("Create"))
-                {
-                    isBusy = true;
-
-                    Client.CreateRepository(new NewRepository(repoName) {
-                        Private = togglePrivate,
-                    }, (repository, ex) => {
-                        Logger.Trace("Create Repository Callback");
-
-                        if (ex != null)
-                        {
-                            error = ex.Message;
-                            isBusy = false;
-                            return;
-                        }
-
-                        if (repository == null)
-                        {
-                            Logger.Warning("Returned Repository is null");
-                            isBusy = false;
-                            return;
-                        }
-
-                        GitClient.RemoteAdd("origin", repository.CloneUrl)
-                                 .Then(GitClient.Push("origin", Repository.CurrentBranch))
-                                 .ThenInUI(Close)
-                                 .Start();
-                    }, owners[selectedOrg] == username ? null : owners[selectedOrg]);
-                }
-                GUI.enabled = true;
-            }
-            GUILayout.EndHorizontal();
-            GUILayout.Space(10);
+        public override void Finish(bool result)
+        {
+            Close();
+            base.Finish(result);
         }
     }
 }
